refactor: :recycle: move global types to global types file
