+++ src/sqljs/tests/sortKeys.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { sortKeys } from '../sibylLib'
+import { sortKeys } from '../../sibylLib'
 
 interface TableRow {
   id: number
