--- src/sqljs/types.ts
@@ -1,13 +1,0 @@
-export interface UpdateArgs<T, K extends string | number | symbol = 'id'> {
-  where: Partial<T>
-  updates: Partial<Omit<T, | K>>
-}
-
-export interface DeleteArgs<T> {
-  where: Partial<T>
-}
-
-export interface DataStructure {
-  columns: string[]
-  values: any[][]
-}
