+++ src/types.ts
@@ -47,3 +47,17 @@ export interface SelectArgs<T> {
   limit?: number
   sort?: Sort<Partial<T>>
 }
+
+export interface UpdateArgs<T, K extends string | number | symbol = 'id'> {
+  where: Partial<T>
+  updates: Partial<Omit<T, | K>>
+}
+
+export interface DeleteArgs<T> {
+  where: Partial<T>
+}
+
+export interface DataStructure {
+  columns: string[]
+  values: any[][]
+}
